% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-inf.R
\name{as_input_file}
\alias{as_input_file}
\title{Convert data.frame to list of inf class object}
\usage{
as_input_file(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Convert data.frame to list of inf class object
}
