% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{bash_extract_file_from_tar}
\alias{bash_extract_file_from_tar}
\title{Extract files from arachive file (.tar)}
\usage{
bash_extract_file_from_tar(archive_file)
}
\arguments{
\item{archive_file}{.tar file path.}
}
\description{
Extract files from arachive file (.tar)
}
