% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{bash_merge_fastq}
\alias{bash_merge_fastq}
\title{Merge some fastq files to a fastq file}
\usage{
bash_merge_fastq(fpath_in_fastq_group, fpath_out_fastq)
}
\arguments{
\item{fpath_in_fastq_group}{.tar file path}

\item{fpath_out_fastq}{.tar file path}
}
\description{
Merge some fastq files to a fastq file
}
