% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rpkm.R
\name{calc_rpkm_from_featurecounts}
\alias{calc_rpkm_from_featurecounts}
\title{Calculate RPKM from featureCounts output}
\usage{
calc_rpkm_from_featurecounts(df)
}
\arguments{
\item{df}{data.frame output from merge_featurecount_output()}
}
\description{
Calculate RPKM from featureCounts output
}
