% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rpkm.R
\name{calc_rpm}
\alias{calc_rpm}
\title{Calculate Reads Per Million mapped reads (RPM)}
\usage{
calc_rpm(readcount)
}
\arguments{
\item{readcount}{read count. atomic vector}
}
\description{
Calculate Reads Per Million mapped reads (RPM)
}
