% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rpkm.R
\name{calc_rpm_from_featurecounts}
\alias{calc_rpm_from_featurecounts}
\title{Calculate RPM from featureCounts output}
\usage{
calc_rpm_from_featurecounts(df)
}
\arguments{
\item{df}{data.frame output from merge_featurecount_output()}
}
\description{
Calculate RPM from featureCounts output
}
