% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rpkm.R
\name{calc_tpm}
\alias{calc_tpm}
\title{Calculate TPM}
\usage{
calc_tpm(readcount, len)
}
\arguments{
\item{readcount}{read count. atomic vector}

\item{len}{feature length. atomic vector}
}
\description{
Calculate TPM
}
