% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_fastp.R
\name{cmd_set_fastp_qf}
\alias{cmd_set_fastp_qf}
\title{Filter fastq file by quality}
\usage{
cmd_set_fastp_qf(inf_list, out_dir = "fastq_qf")
}
\arguments{
\item{out_dir}{A path of directory for saving output files}
}
\description{
Filter fastq file by quality
}
