% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cmdopt.R
\name{cmd_set_merge_gz}
\alias{cmd_set_merge_gz}
\title{Merge some fastq files to a fastq file}
\usage{
cmd_set_merge_gz(gziped_files, gzip_out)
}
\arguments{
\item{gziped_files}{.tar file path}

\item{gzip_out}{.tar file path}
}
\description{
Merge some fastq files to a fastq file
}
