% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc_data}
\alias{fastqc_data}
\title{Extract information from the FastQC data.txt file}
\usage{
fastqc_data(path)
}
\arguments{
\item{path}{A file path to the FastQC data.txt}
}
\description{
Extract information from the FastQC data.txt file
}
