% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc_summary}
\alias{fastqc_summary}
\title{Summarize FastQC summary.txt files}
\usage{
fastqc_summary(paths)
}
\arguments{
\item{paths}{character vector which contains file paths to the FastQC summary.txt}
}
\description{
Summarize FastQC summary.txt files
}
