% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools-stat.R
\name{glue_bam2wig}
\alias{glue_bam2wig}
\title{bam2wig.py}
\usage{
glue_bam2wig(head_label, in_dir, chr_name_len_txt, out_dir = in_dir)
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{chr_name_len_txt}{text file. chromatin name and length}

\item{out_dir}{A path of directory for saving output files}
}
\description{
bam2wig.py
}
