% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools-stat.R
\name{glue_bam_index}
\alias{glue_bam_index}
\title{samtools index}
\usage{
glue_bam_index(head_label, in_dir)
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}
}
\description{
samtools index
}
