% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools-stat.R
\name{glue_bam_sort}
\alias{glue_bam_sort}
\title{samtools sort}
\usage{
glue_bam_sort(head_label, in_dir, out_dir = in_dir)
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}
}
\description{
samtools sort
}
