% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools-stat.R
\name{glue_bamstat}
\alias{glue_bamstat}
\title{Create fastqc reports}
\usage{
glue_bamstat(head_label, in_dir, out_dir = in_dir)
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}
}
\description{
Create fastqc reports
}
