% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-bowtie2.R
\name{glue_bowtie2_rm_contami_pe}
\alias{glue_bowtie2_rm_contami_pe}
\title{Remove paired end reads from contamination with bowtie2}
\usage{
glue_bowtie2_rm_contami_pe(head_label, in_dir, fasta_path,
  out_dir = "fastq_rm", idx_dir = "./idx_bw2", core_num = 4)
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{fasta_path}{.tar file path}

\item{out_dir}{A path of directory for saving output files}

\item{idx_dir}{directory path for the indexing output}

\item{core_num}{number of cpu core}
}
\description{
Remove paired end reads from contamination with bowtie2
}
