% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-bowtie.R
\name{glue_bowtie_bamsort}
\alias{glue_bowtie_bamsort}
\title{Map with bowtie and output a sorted bam file}
\usage{
glue_bowtie_bamsort(head_label, index_name, in_dir = "./fastq",
  out_dir = "./mapped_by_bowtie", core_num = 2)
}
\arguments{
\item{head_label}{.tar file path}

\item{index_name}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}

\item{core_num}{number of threads to use}
}
\description{
Map with bowtie and output a sorted bam file
}
