% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-bowtie.R
\name{glue_bowtie_genome_generate}
\alias{glue_bowtie_genome_generate}
\title{Generate genome index for bowtie}
\usage{
glue_bowtie_genome_generate(fasta_path, index_name)
}
\arguments{
\item{fasta_path}{fasta file containing reference sequences to be aligned for}

\item{index_name}{Name of the index}
}
\description{
Generate genome index for bowtie
}
