% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools-stat.R
\name{glue_check_gz_size}
\alias{glue_check_gz_size}
\title{Check the decompressed file size of gzipped file}
\usage{
glue_check_gz_size(in_dir)
}
\arguments{
\item{in_dir}{A path of directory which contains input files}
}
\description{
Check the decompressed file size of gzipped file
}
