% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cuff.R
\name{glue_cuffcompare}
\alias{glue_cuffcompare}
\title{cuffcompare commands}
\usage{
glue_cuffcompare(merged_gtf_path = "./cuff/2_merge/merged.gtf",
  out_dir = "./cuff/3_merge",
  ref_gtf_path = "./TAIR10_GFF3_genes_transposons.gtf", use_RABT = T)
}
\arguments{
\item{merged_gtf_path}{.tar file path}

\item{out_dir}{A path of directory for saving output files}

\item{ref_gtf_path}{.tar file path}

\item{use_RABT}{.tar file path}
}
\description{
cuffcompare commands
}
