% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_cufflinks}
\alias{glue_cufflinks}
\title{Calculate FPKM by Cufflinks}
\usage{
glue_cufflinks(head_label, in_dir, out_dir = "./fpkm", core_num = 2,
  gff_path = "./TAIR10_GFF3_genes_transposons.gff")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}

\item{core_num}{.tar file path}

\item{gff_path}{.tar file path}
}
\description{
Calculate FPKM by Cufflinks
}
