% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cuff.R
\name{glue_cuffmerge}
\alias{glue_cuffmerge}
\title{cuffmerge commands}
\usage{
glue_cuffmerge(cuff_gtf_dir = "./cuff/1_links",
  out_dir = "./cuff/2_merge", core_num = 2,
  ref_gtf_path = "./TAIR10_GFF3_genes_transposons.gtf")
}
\arguments{
\item{cuff_gtf_dir}{.tar file path}

\item{out_dir}{A path of directory for saving output files}

\item{core_num}{number of cpu core}

\item{ref_gtf_path}{.tar file path}
}
\description{
cuffmerge commands
}
