% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cuff.R
\name{glue_cuffnorm}
\alias{glue_cuffnorm}
\title{cuffnorm commands}
\usage{
glue_cuffnorm(head_labels, in_dir = "./cuff/4_quant",
  out_dir = "./cuff/6_norm",
  merged_gtf_path = "./cuff/2_merge/merged.gtf", core_num = 2)
}
\arguments{
\item{head_labels}{.tar file path}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{merged_gtf_path}{.tar file path}

\item{core_num}{number of cpu core}
}
\description{
cuffnorm commands
}
