% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cuff.R
\name{glue_cuffquant}
\alias{glue_cuffquant}
\title{cuffquant commands}
\usage{
glue_cuffquant(head_label, in_dir, out_dir = "./cuff/4_quant",
  core_num = 2, merged_gtf_path = "./cuff/2_merge/merged.gtf")
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{core_num}{number of cpu core}

\item{merged_gtf_path}{.tar file path}
}
\description{
cuffquant commands
}
