% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cutadapt.R
\name{glue_cutadapt_singleend}
\alias{glue_cutadapt_singleend}
\title{Clipping adaptor from single end reads}
\usage{
glue_cutadapt_singleend(label, adaptor, in_dir = "./fastq",
  out_dir = "./fastq")
}
\arguments{
\item{label}{.tar file path}

\item{adaptor}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Clipping adaptor from single end reads
}
