% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_fastp.R
\name{glue_fastp_qf}
\alias{glue_fastp_qf}
\title{Filter fastq file by quality}
\usage{
glue_fastp_qf(head_label, in_dir = "./fastq", out_dir = "./fastq_qf")
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}
}
\description{
Filter fastq file by quality
}
