% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-fastx-toolkit.R
\name{glue_fastq_qf_cl}
\alias{glue_fastq_qf_cl}
\title{Filter fastq file by quality and Trim adaptor sequence}
\usage{
glue_fastq_qf_cl(head_label, in_dir = "./fastq",
  out_dir = "./fastq_qf_cl")
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}
}
\description{
Filter fastq file by quality and Trim adaptor sequence
}
