% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_featurecounts}
\alias{glue_featurecounts}
\title{featureCounts}
\usage{
glue_featurecounts(head_label, in_dir, out_dir = "./readcount",
  gtf_path = "./TAIR10_GFF3_genes_transposons.gtf")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}

\item{gtf_path}{.tar file path}
}
\description{
featureCounts
}
