% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools-stat.R
\name{glue_fqgz2rfqxz}
\alias{glue_fqgz2rfqxz}
\title{convert fastq.gz to rfq.xz}
\usage{
glue_fqgz2rfqxz(head_label, in_dir, out_dir = in_dir, core_num = 4)
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{core_num}{number of cpu core}
}
\description{
convert fastq.gz to rfq.xz
}
