% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_genomecov}
\alias{glue_genomecov}
\title{Calculate genome coverage}
\usage{
glue_genomecov(head_label, in_dir = ".", out_dir = ".")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Calculate genome coverage
}
