% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_gff2bed}
\alias{glue_gff2bed}
\title{Convert gff3 file to bed file}
\usage{
glue_gff2bed(head_label, in_dir = ".", out_dir = ".")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Convert gff3 file to bed file
}
