% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-hisat.R
\name{glue_hisat_bamsort}
\alias{glue_hisat_bamsort}
\title{Map with STAR and output a sorted bam file}
\usage{
glue_hisat_bamsort(read1, read2, core_num = 2, index_name = "TAIR10",
  out = "hoge")
}
\arguments{
\item{read1}{.tar file path}

\item{read2}{.tar file path}

\item{core_num}{.tar file path}

\item{index_name}{.tar file path}

\item{out}{.tar file path}
}
\description{
Map with STAR and output a sorted bam file
}
