% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-hisat.R
\name{glue_hisat_genome_generate}
\alias{glue_hisat_genome_generate}
\title{Generate genome index for HISAT2}
\usage{
glue_hisat_genome_generate(fasta_path = "./allChr.fas",
  idx_name = "TAIR10", idx_dir = "./idx_hisat", core_num = 2)
}
\arguments{
\item{fasta_path}{genome fasta file path for indexing}

\item{idx_name}{index name}

\item{idx_dir}{directory path for the indexing output}

\item{core_num}{number of cpu core}
}
\description{
Generate genome index for HISAT2
}
