% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-hisat.R
\name{glue_pe_hisat_bamsort}
\alias{glue_pe_hisat_bamsort}
\title{Map paired-end reads with hisat and output a sorted bam file}
\usage{
glue_pe_hisat_bamsort(head_label, core_num = 2, idx_name = "TAIR10",
  idx_dir = "./idx_hisat", in_dir = "./fastq",
  out_dir = "./mapped_by_hisat")
}
\arguments{
\item{head_label}{sample label}

\item{core_num}{number of cpu core}

\item{idx_name}{index name}

\item{idx_dir}{directory path for the indexing output}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}
}
\description{
Map paired-end reads with hisat and output a sorted bam file
}
