% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools-stat.R
\name{glue_rfqxz2fqgz}
\alias{glue_rfqxz2fqgz}
\title{convert rqf.gz to  fastq.gz}
\usage{
glue_rfqxz2fqgz(head_label, in_dir, out_dir = in_dir, core_num = 4)
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{core_num}{number of cpu core}
}
\description{
convert rqf.gz to  fastq.gz
}
