% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cutadapt.R
\name{glue_se_cutadapt}
\alias{glue_se_cutadapt}
\title{Clipping adaptor from single end reads}
\usage{
glue_se_cutadapt(head_label, adaptor, in_dir = "./fastq",
  out_dir = "./fastq")
}
\arguments{
\item{head_label}{sample label}

\item{adaptor}{.tar file path}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}
}
\description{
Clipping adaptor from single end reads
}
