% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-featurecounts.R
\name{glue_se_featurecounts}
\alias{glue_se_featurecounts}
\title{featureCounts}
\usage{
glue_se_featurecounts(head_label, in_dir, out_dir = "./readcount",
  core_num = 2, gtf_path = "./TAIR10_GFF3_genes_transposons.gtf",
  feat_type = "exon")
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{core_num}{number of cpu core}

\item{gtf_path}{.tar file path}

\item{feat_type}{specify feature type}
}
\description{
featureCounts
}
