% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-hisat.R
\name{glue_se_hisat_bamsort}
\alias{glue_se_hisat_bamsort}
\title{Map single-end reads with hisat and output a sorted bam file}
\usage{
glue_se_hisat_bamsort(head_label, core_num = 2, index_name = "TAIR10",
  index_dir = "./genomeidx_hisat", in_dir = "./fastq",
  out_dir = "./mapped_by_hisat")
}
\arguments{
\item{head_label}{.tar file path}

\item{core_num}{.tar file path}

\item{index_name}{.tar file path}

\item{index_dir}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Map single-end reads with hisat and output a sorted bam file
}
