% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_star_bamsort}
\alias{glue_star_bamsort}
\title{Map with bowtie and output a sorted bam file}
\usage{
glue_star_bamsort(head_label, in_dir = "./fastq",
  out_dir = "./mapped_by_star", stargenome_dir = "./stargenome",
  gff_path = "TAIR10_GFF3_genes_transposons.gff")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}

\item{stargenome_dir}{.tar file path}

\item{gff_path}{.tar file path}
}
\description{
Map with bowtie and output a sorted bam file
}
