% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-star.R
\name{glue_star_genome_generate}
\alias{glue_star_genome_generate}
\title{Generate genome index for STAR}
\usage{
glue_star_genome_generate(stargenome_dir = "./stargenome",
  chr_fasta_path = "./allChr.fas",
  gff_path = "./TAIR10_GFF3_genes_transposons.gff", core_num = 2)
}
\arguments{
\item{stargenome_dir}{.tar file path}

\item{chr_fasta_path}{.tar file path}

\item{gff_path}{.tar file path}

\item{core_num}{.tar file path}
}
\description{
Generate genome index for STAR
}
