% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-featurecounts.R
\name{merge_featurecount_output}
\alias{merge_featurecount_output}
\title{Merging read count tsv file from featureCounts}
\usage{
merge_featurecount_output(in_dir = "./readcount")
}
\arguments{
\item{in_dir}{.tar file path}
}
\description{
Merging read count tsv file from featureCounts
}
