% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cmdopt.R
\name{new_cmdopt}
\alias{new_cmdopt}
\title{Create new cmdopt class object}
\usage{
new_cmdopt(..., cmd_id = NULL)
}
\arguments{
\item{...}{...}

\item{cmd_id}{command type specifier}
}
\description{
Create new cmdopt class object
}
