% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nothing.R
\name{param_general}
\alias{param_general}
\title{Parameters for general use}
\usage{
param_general(in_dir, out_dir, head_label, core_num)
}
\arguments{
\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{head_label}{sample label}

\item{core_num}{number of cpu core}
}
\description{
Parameters for general use
}
