% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nothing.R
\name{param_genomeidx}
\alias{param_genomeidx}
\title{Parameters for genome indexing commands}
\usage{
param_genomeidx(fasta_path, idx_name, idx_dir)
}
\arguments{
\item{fasta_path}{genome fasta file path for indexing}

\item{idx_name}{index name}

\item{idx_dir}{directory path for the indexing output}
}
\description{
Parameters for genome indexing commands
}
