% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cuff.R
\name{pipeline_cuffdiff}
\alias{pipeline_cuffdiff}
\title{pipeline for cufflinks}
\usage{
pipeline_cuffdiff(head_labels, in_dir, core_num = 2,
  ref_gtf_path = "./TAIR10_GFF3_genes_transposons.gtf")
}
\arguments{
\item{head_labels}{.tar file path}

\item{in_dir}{A path of directory which contains input files}

\item{core_num}{number of cpu core}

\item{ref_gtf_path}{.tar file path}
}
\description{
pipeline for cufflinks
}
