% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_ext.R
\name{set_file_ext}
\alias{set_file_ext}
\title{Set file extensions}
\usage{
set_file_ext(fastq_ext = "fq", bam_ext = "bam", bigwiggle_ext = "bw")
}
\arguments{
\item{fastq_ext}{fastq file extension}

\item{bam_ext}{bam file extension}

\item{bigwiggle_ext}{bigwiggle file extension}
}
\description{
Set file extensions
}
