% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rpkm.R
\name{calc_rpkm}
\alias{calc_rpkm}
\title{Calculate Reads Per Killobase of exon per Million mappled reads (RPKM)}
\usage{
calc_rpkm(readcount, len)
}
\arguments{
\item{readcount}{read count. atomic vector}

\item{len}{feature length. atomic vector}
}
\description{
Calculate Reads Per Killobase of exon per Million mappled reads (RPKM)
}
