% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-bowtie2.R
\name{glue_bowtie2_bamsort}
\alias{glue_bowtie2_bamsort}
\title{Map with bowtie2 and output a sorted bam file}
\usage{
glue_bowtie2_bamsort(head_label, index_name, core_num = 2,
  in_dir = "./fastq", out_dir = "./mapped_by_bowtie2")
}
\arguments{
\item{head_label}{.tar file path}

\item{index_name}{.tar file path}

\item{core_num}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Map with bowtie2 and output a sorted bam file
}
