% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_bowtie_bamsort}
\alias{glue_bowtie_bamsort}
\title{Map with bowtie and output a sorted bam file}
\usage{
glue_bowtie_bamsort(head_label, in_dir = "./fastq",
  out_dir = "./mapped_by_bowtie")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Map with bowtie and output a sorted bam file
}
