% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_fastqc}
\alias{glue_fastqc}
\title{Create fastqc reports}
\usage{
glue_fastqc(in_dir = "./fastq", out_dir = "./fastqc_out",
  core_num = 4)
}
\arguments{
\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{core_num}{number of cpu core}
}
\description{
Create fastqc reports
}
