% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_fastqc}
\alias{glue_fastqc}
\title{Create fastqc reports}
\usage{
glue_fastqc(in_dir = "./fastq", out_dir = "./fastqc_out")
}
\arguments{
\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Create fastqc reports
}
