% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_genomecov}
\alias{glue_genomecov}
\title{Calculate genome coverage}
\usage{
glue_genomecov(fpath_sorted_bam, fpath_out)
}
\arguments{
\item{fpath_sorted_bam}{.tar file path}

\item{fpath_out}{.tar file path}
}
\description{
Calculate genome coverage
}
