% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-hisat.R
\name{glue_hisat_genome_generate}
\alias{glue_hisat_genome_generate}
\title{Generate genome index for HISAT2}
\usage{
glue_hisat_genome_generate(chr_fasta_path = "./allChr.fas",
  index_name = "TAIR10")
}
\arguments{
\item{chr_fasta_path}{.tar file path}

\item{index_name}{.tar file path}
}
\description{
Generate genome index for HISAT2
}
