% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-cutadapt.R
\name{glue_pe_cutadapt}
\alias{glue_pe_cutadapt}
\title{Clipping adaptor from pair end reads}
\usage{
glue_pe_cutadapt(label, f_adaptor, r_adaptor, overlap = 3,
  minimum_length = 20, in_dir = "./fastq", out_dir = "./fastq")
}
\arguments{
\item{label}{.tar file path}

\item{f_adaptor}{.tar file path}

\item{r_adaptor}{.tar file path}

\item{overlap}{.tar file path}

\item{minimum_length}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}
}
\description{
Clipping adaptor from pair end reads
}
