% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-featurecounts.R
\name{glue_se_featurecounts}
\alias{glue_se_featurecounts}
\title{featureCounts}
\usage{
glue_se_featurecounts(head_label, in_dir, out_dir = "./readcount",
  core_num = 2, gtf_path = "./TAIR10_GFF3_genes_transposons.gtf")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{out_dir}{.tar file path}

\item{core_num}{.tar file path}

\item{gtf_path}{.tar file path}
}
\description{
featureCounts
}
