% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-star.R
\name{glue_se_star_bamsort}
\alias{glue_se_star_bamsort}
\title{Map with STAR and output a sorted bam file}
\usage{
glue_se_star_bamsort(head_label, in_dir = "./fastq",
  out_dir = "./mapped_by_star", idx_dir = "./idx_star",
  gff_path = "TAIR10_GFF3_genes_transposons.gff")
}
\arguments{
\item{head_label}{sample label}

\item{in_dir}{A path of directory which contains input files}

\item{out_dir}{A path of directory for saving output files}

\item{idx_dir}{directory path for the indexing output}

\item{gff_path}{.tar file path}
}
\description{
Map with STAR and output a sorted bam file
}
