% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_star_pipeline}
\alias{glue_star_pipeline}
\title{Calculate FPKM by Cufflinks}
\usage{
glue_star_pipeline(head_label, in_dir = "./fastq",
  stargenome_dir = "./stargenome",
  gff_path = "TAIR10_GFF3_genes_transposons.gff")
}
\arguments{
\item{head_label}{.tar file path}

\item{in_dir}{.tar file path}

\item{stargenome_dir}{.tar file path}

\item{gff_path}{.tar file path}
}
\description{
Calculate FPKM by Cufflinks
}
